+++ car/src/main/java/com/nav/agri/repositories/StockRecordRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockRecordRepository extends JpaRepository<StockRecord, Long> {
+public interface StockRecordRepository extends JpaRepository<StockRecord, Integer> {
 }
